+++ src/main/java/net/thedragonteam/armorplus/ArmorPlus.java
@@ -42,7 +42,7 @@ public class ArmorPlus {
      * Updates every time a new block, item or features is added or change,
      * resets on MAJOR changes
      */
-    public static final int MINOR = 1;
+    public static final int MINOR = 2;
     /**
      * Updates every time a bug is fixed or issue solved or very minor code changes,
      * resets on MINOR changes