+++ src/main/java/net/thedragonteam/armorplus/compat/jei/ultitechbench/UltiTechBenchShapelessOreRecipeWrapper.java
@@ -22,14 +22,12 @@ public class UltiTechBenchShapelessOreRecipeWrapper extends AbstractShapelessRec
     public UltiTechBenchShapelessOreRecipeWrapper(IGuiHelper guiHelper, ShapelessOreRecipe recipe) {
         super(guiHelper);
         this.recipe = recipe;
-        for (Object input : this.recipe.getInput()) {
-            if (input instanceof ItemStack) {
-                ItemStack itemStack = (ItemStack) input;
-                if (itemStack.stackSize != 1) {
-                    itemStack.stackSize = 1;
-                }
+        this.recipe.getInput().stream().filter(input -> input instanceof ItemStack).forEachOrdered(input -> {
+            ItemStack itemStack = (ItemStack) input;
+            if (itemStack.stackSize != 1) {
+                itemStack.stackSize = 1;
             }
-        }
+        });
     }
 
     @Override